+++ app/src/main/java/roboyard/eclabs/GridGameScreen.java
@@ -33,6 +33,7 @@ public class GridGameScreen extends GameScreen {
     private int solutionMoves = 0; // store the optimal number of moves for the solution globally
     private int NumDifferentSolutionsFound = 0; // store the number of different solution globally
     private int numSolutionClicks = 0; // count how often you clicked on the solution button, each time the shown count goes down by one
+    private int numDifferentSolutionClicks = 0; // count how often you clicked on the solution button again to show a different solution
     private int showSolutionAtHint = 5; // interval between the first hint and the current optimal solution (will be set to random 3..5 later
 
     private static int goodPuzzleMinMoves = 8; // below this number of moves there is a special hint shown from the start
@@ -386,7 +387,7 @@ public class GridGameScreen extends GameScreen {
         {
             isSolved = true;
             buttonSolve.setEnabled(true);
-            GameSolution solution = solver.getSolution();
+            GameSolution solution = solver.getSolution(numDifferentSolutionClicks);
             solutionMoves=0;
             for(IGameMove m : solution.getMoves()){
                 solutionMoves++;
@@ -771,8 +772,17 @@ public class GridGameScreen extends GameScreen {
     private class ButtonSolution implements IExecutor{
         public void execute(){
             if(numSolutionClicks >= showSolutionAtHint) {
-                GameSolution solution = solver.getSolution();
+                GameSolution solution = solver.getSolution(numDifferentSolutionClicks);
                 showSolution(solution);
+                if(NumDifferentSolutionsFound > 1){
+                    // if solution is shown and there are more than 1 solutions found:
+                    if(numDifferentSolutionClicks >= NumDifferentSolutionsFound-1){
+                        numDifferentSolutionClicks = 0;
+                    }else{
+                        numDifferentSolutionClicks++;
+                    }
+                    gameManager.requestToast("Press again to see solution " + (numDifferentSolutionClicks + 1), false);
+                }
             }else{
                 numSolutionClicks++;
                 if(numSolutionClicks < showSolutionAtHint) {
@@ -780,8 +790,6 @@ public class GridGameScreen extends GameScreen {
                 } else {
                     gameManager.requestToast("Press again to see the solution.", false);
                 }
-                // TODO: if solution is shown and there are more than 1 solutions found:
-                //       show toast "Press again to se an alternative solution"
             }
         }
     }