+++ app/src/main/java/roboyard/eclabs/solver/SolverDD.java
@@ -23,14 +23,12 @@ public class SolverDD implements ISolver{
 
     private SolverStatus solverStatus;
     private Solver solver;
-    private Solution solution;
     private List<Solution> solutions;
     private final RRPiece[] pieces;
 
     public SolverDD(){
         solver = null;
         solverStatus = SolverStatus.idle;
-        solution = null;
         solutions = null;
         pieces = new RRPiece[4];
     }
@@ -53,7 +51,7 @@ public class SolverDD implements ISolver{
         try {
             solutions = solver.execute();
             if(solutions.size() != 0){
-                solution = solutions.get(0);
+                Solution solution = solutions.get(0);
                 System.out.println(solution.toString());
                 solverStatus = SolverStatus.solved;
             }else{
@@ -72,9 +70,14 @@ public class SolverDD implements ISolver{
         return this.solutions;
     }
 
-    public GameSolution getSolution(){
+    /**
+     * get the solution number num from the list of found different solutions and add all moves to the result
+     * @param num number of the solution in the solutions list
+     * @return GameSolution with all moves in that solution
+     */
+    public GameSolution getSolution(int num){
         GameSolution s = new GameSolution();
-
+        Solution solution = solutions.get(num);
         solution.resetMoves();
         Move m = solution.getNextMove();
         while (m != null){